Move provider registry into plugins service
New interface location:

- `services/plugins/registry.ProviderMap[T]` owns the thread-safe named provider collection used by plugin invocation, agent, workflow, MCP, and bootstrap wiring.
- `services/plugins/registry.Registry` continues to group `core.AuthenticationProvider` and `core.Provider` registries.

Usage examples:

```go
providers := registry.New().Providers
if err := providers.Register("slack", provider); err != nil {
    return err
}
```

```go
broker := invocation.NewBroker(&providers, users, externalCredentials)
```

This removes service imports of `internal/registry` and places provider registry primitives under the plugins service boundary.
