+++ gestaltd/internal/bootstrap/cleanup.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/valon-technologies/gestalt/server/core"
 	"github.com/valon-technologies/gestalt/server/internal/providerdev"
-	"github.com/valon-technologies/gestalt/server/internal/registry"
+	"github.com/valon-technologies/gestalt/server/services/plugins/registry"
 )
 
 // CloseProviders closes all registered providers that implement io.Closer.
