+++ gestaltd/internal/bootstrap/result_boundary_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/valon-technologies/gestalt/server/core"
 	coretesting "github.com/valon-technologies/gestalt/server/core/testing"
 	"github.com/valon-technologies/gestalt/server/internal/bootstrap"
-	"github.com/valon-technologies/gestalt/server/internal/registry"
+	"github.com/valon-technologies/gestalt/server/services/plugins/registry"
 )
 
 type closableProvider struct {
